Db: fix unsynchronized position calc from amplitudes change
Now, on processing the Monitoring Amplitude A, it forces the processing of
all other amplitudes and position calculation. Also, the first record has
its MEDL value changed to -1, so even if its value has not changed it will
still "send" its value to clients. This is the expected behavior, as the
monitoring amplitudes might not change, but we still want to have the position
calculated.