cfg: use stronger coupling for bpm-sw@ units
What we want is that when the respective
bpm-sw@ instance unit is started, the corresponding
IOCs (BE and FE - bpm-sw-be-ioc@ and bpm-sw-fe-ioc@
instance units) get started as well.

Moreover, when the bpm-sw@ instance unit disappears
or is stopped we also want the IOCs are stopped.
This is because the IOCs need to communicate
to the bpm-sw services to to function.