BPMApp: use asyn Addresses to select between trigger channels
The maximum number of waveforms must be at least
the number of triggers available on the FPGA. Although
this is used to alloc the number of waveforms, it's not
used by getAddress () by the NDArray plugins, as this function
returns the address that is declared on plugin startup
(NDStdArraysConfigure function, NDArrayAddr). So, we
are free to use all of the addresses that are set by
the database. In summary, we use the different addresses
to call different trigger channel functions.