BPMApp: only send abort event if we are reading
This is important as the events are "sticky".
Even if there are no threads waiting, the next
time the thread calls epicsEventWait, it will
receive the event.

This avoids the case where we set the "abort"
trigger on a IDLE state. This would cause the next
acquisition to abort, which is not the expected
behavior.