opi: fix abort/start acquisition each time the panel was opened
Every time an acquisition panel was opened
it aborted the current acquisition and started
a next one. This was due to the fact that
the script was being triggered by any PV
event, even for the initial value.

This was fixed by checking the box
"don't trigger script on PV first value".