BPMApp: change PM acquisition to TBT rate
This is mroe suitable for this kind of event.
Also, if we run at the ADC rate, we use
roughly 1600MB/s (2 (bytes per sample) *
4 (ADC channels) * ~200e6 (ADC rate in MHz).

This is half the bandwidth of the DDR3.
It seems ok, but we share this same bandwidth
with another BPM. So, we would end up using
all of the availabe andwidth only for Post-Mortem.