BPMApp: refactor Single Pass task to correct bahavior
The Single Pass Task was not getting created
in the first place. Also, some expected behavior,
like stoppting/aborting the Single Pass acquisition
was not being followed. So, we fixed them here.

This is a big commit that should be split into
multiple ones, but the changes are intricate
that a single one of them wouldn't be much of
use.