BPMApp: only abort acquistiion in HW if necessary
As MultiBunch and SinglePass uses the same
Acquisition Core in HW, we can't abort
the core everytime we set the trigger
(every acquisition). Doing so, might cause
inconsistencies, as we might abort the core
and the condition
    (!readingActive[bpmMode][addr] &&
     !repetitiveTrigger[bpmMode][addr])
might be false (readingActive still true)
causing the acqTask to be stuck (as the core
was aborted the FSM returned to the IDLE
status, never signalling a completed
acquistion) until we do an actual abort
operation.

This fixes #9 github issue.