BPMApp: add new ActiveClkRstAdcs record
This is important if we are using external
FMC clock and it becomes on only after the
board is initialized. As the AD9510 and ADCs
needs a clock to be configured, we will
end up in a uninitialized state.

So, for now, the user must reset the ADCs
and AD9510 manually.