BPMApp: change constructor initiallization order
This was done so we could read HW parameters
after setting some default in parameter library.

Now, we do it like this:

1) Create the parameters in the paramter library
2) Set the function hash table
3) Set some defaults in parameter library

We could read the paramters from hardware as a
set 4, but ASYN default device support for Int32 and
Digital32 already do this "bumpless restore" by default.
So, on record initialization, it read from hardware.

This is explained by Mark Rivers in this tech talk
(http://www.aps.anl.gov/epics/tech-talk/2013/msg01466.php):

Subject: 	RE: Initial Value of a BO Record From: 	Mark
Rivers <rivers@cars.uchicago.edu> To: 	"'Zenon Szalata'"
<zms@slac.stanford.edu>, tech-talk <tech-talk@aps.anl.gov>
Date: 	Thu, 11 Jul 2013 16:44:55 +0000

Hi Zen,

The concept of the "bumpless reboot" is as follows.

- The init_record function in the asyn device support for
the asynInt32 interface attempts to read the value from the
driver.  This happens during iocInit.  - If the driver
returns asynSuccess then the value returned by the driver is
placed into the RVAL field in the bo record, and it tells
the record to convert the RVAL value to the VAL field.  This
will make the record agree with the hardware after iocInit.
- If the driver does not return asynSuccess then the device
support does not change the RVAL field of the bo record, and
returns 2 (do not convert).  - If you are using
asynPortDriver and you are implementing the readInt32 method
then you have control of whether to return asynSuccess or
asynError when called during iocInit.  You may need to have
a flag indicating whether the value has ever been written by
a call to writeInt32, or look at the value of the EPICS
global "interruptAccept", etc.  - If you are using
asynPortDriver and you are NOT implementing the readInt32
method then the asynPortDriver::readInt32 base class
implementation will be used.  It returns asynSuccess if a
value has ever been written to the parameter library with
setIntegerParam() for that parameter and address.  If
setIntegerParam has not yet been called then it returns
asynError.