opi: replace local UpdateTime PV to real PV
This was just a precaution to the user
so the UpdateTime was set before changing
the trigger. But, in fact, it's better
to let the user change this at will as
it gives a more pleasant UI response.

Also, the necessary handling of changing
this value in the middle of an acquisition
is already treater in IOC, so no special
handling is needed.

This fixes #13 github issue