opi: remove "stop" MONIT button for simplicity
Now, all the behavior is controllable via
Update Rate button and the associated
combo box.

This also, fixes the issue that sometimes
we will end up with a stopped monitoring
rate and sometime the correct rate.

For some reason, both buttons (start/stop)
were being executed at the same time and the
last one will be the effective one.