BPMApp: improve MONIT. AMP/POS responsiveness by using a new thread
Previously epics SCAN 0.1 thread was not suitable for
the MONIT task for 2 reasons.

1) It had a fixed scan rate of 0.1Hz which did not matches
the MONIT. rate, which is ~10Hz (9, 10, 11, depending on the
decimation rate). This would cause inconsistent sampling and
eventually a missing sample.

2) It had a high frequency of "hipcups", in the sense
that very often (once every 1-2 seconds) it delayed the
sapling of a new MONIT AMP/POS and update the PV twice
at the sme timestamp.

Now, we spawn a new high priority task that keeps
sampling the MONIT. FIFO in roder to check if there
are new data. If yes, it fetches the data and update the
PV/timestamp. If not, nothing happens.

For this to operate reliably, we need to poll at a higher
rate than the sample rate ~10Hz) so as to not miss any sample.
So, we set a default poll rate of 20Hz.