opi: update scan rate meaning
Now, we can only poll the MONIT data at the
specified rate, but this is limited by the
actual valid data rate in the FPGA. So, we just
need to be faster than ~10Hz so to not loose
any sample.

This poll rate should only be changed for debugging
purposes, as this affects the correctness of acquired
MONIT. data.