configure/RELEASE: group modules that share paths so epics 3.14.15.6 does not complain
3.14.15.6 EPICS release added this check that
can be seen at:
http://www.aps.anl.gov/epics/base/R3-14/12-docs/RELEASE_NOTES.html,
Additional RELEASE file checks.

The problem lied in two parts. Firstly, we have defined
AREA_DETECTOR_COMPAT, so that synApps 5_7 and 5_8
can use the same area detector path. Second, we were,
unnecessarily, using the areaDetector dbd, which redefined
ADCORE. So, the paths would never be grouped together if
we kept using the areaDetector dbd.

The solution was just to remove it and set the EPICS_BASE
at the RELEASE file.