BPMApp: fix processing of *ArrayData records for NELM changes
The *WvfElemSeq record was processing the *ArrayData
records everytime a change in the *UncalArrayData
changed (i.e., at every new acquisition).

This caused, for uncalibrated aquisitions, the final
ArrayData record to be processed at least twice, one
for each NELM change and another when the data actually
changed.

The fix was to just use NPP instead of PP, as the record
will be process anyway at a later time.