BPMApp: remove unlock/lock around doCallbacksGenericPointer ()
As of ADCore-R3-0
(https://github.com/areaDetector/ADCore/blob/R3-0/RELEASE.md):

"
Previously all plugins were releasing the
asynPortDriver lock when calling doCallbacksGenericPointer().
This was based on a very old observation of a deadlock
problem if the the lock was not released. Releasing
the lock causes serious problems with plugins running
multiple threads, and probably was never needed.
Most plugins no longer call doCallbacksGenericPointer()
directly because it is now done in NDPluginDriver::
endProcessCallbacks(). The lock is no longer released
when calling doCallbacksGenericPointer(). The simDetector
driver has also been modified to no longer release the
lock when calling plugins with doCallbacksGenericPointer(),
and all other drivers should be modified as well. It
is not really a problem with drivers however, since
the code doing those callbacks is normally only
running in a single thread.
"