BPMApp: remove TSEL from BPMPos.
This is a workaround for [1]. The timestamp from drvBPM is wrong and
ends up being propagated from the ArrayData records to the final
position ones, which is not desired. This makes it so the final position
records' timestamps are from when they were processed, which is not an
issue for existing clients.

BPMSP (Single Pass) and BPMDsp (used for Monit rate position data)
weren't changed, since the former is no longer used, and timestamps are
relevant for clients of the latter.

[1] https://github.com/lnls-dig/bpm-epics-ioc/issues/79