BPMApp: fix phys trig PVs being out of sync with hw.
This issue was observed during operation after restarting BPM AFCs,
where the readback PVs had the same values as the setpoints, but forcing
the readback PVs to be processed showed that those values hadn't
actually been written to hardware (or had somehow been overwritten).

Since these PVs directly impact the orbit interlock reliability, we
force them to be read from the hardware periodically; and, to avoid
having to wait up to one second for the readback to be updated after a
write to the setpoint, we add a FLNK to the setpoints.